% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ben_dhs06}
\alias{ben_dhs06}
\title{Preprocessed Benin DHS 06 dataset}
\format{
An object of class \code{data.frame} with 89282 rows and 19 columns.
}
\source{
- DHS Program: https://ohss.dhs.gov/about-our-data
 - OPHI do-file: https://ophi.org.uk/global-mpi/2024
}
\usage{
ben_dhs06
}
\description{
The Benin DHS 06 dataset containing preprocessed data using the OPHI do-file from the Demographic and Health Survey (DHS) program.
}
\references{
OPHI and UNDP (2024). Global Multidimensional Poverty Index 2024: Poverty amid conflict. United Nations Development Programme (UNDP) and Oxford Poverty and Human Development Initiative (OPHI), University of Oxford.
}
\author{
OPHI
}
\keyword{datasets}
