% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.mpitb.R
\name{coef.cotframe}
\alias{coef.cotframe}
\title{Extract the coefficients from the estimated cross-sectional measures}
\usage{
\method{coef}{cotframe}(object, ...)
}
\arguments{
\item{object}{a "cotframe"-class object}

\item{...}{other arguments.}
}
\value{
Coefficients extracted from the model \code{lframe} object.
}
\description{
Extract the coefficients from the estimated cross-sectional measures
}
\details{
The \code{coef} method for "cotframe"-class objects find the point estimates
from the changes over time data frame. This method work for
only one measure \code{c("M0","H","A","hd","hdk")}. Then, user should subset the
data frame with the estimates by the chosen measure (including other preferred categories, i.e.,
poverty cut-off, subgroup, etc.)
}
\seealso{
\code{confint}, and \code{summary} methods, and \code{mpitb.est} function.
}
\author{
Ignacio Girela
}
