% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.mpitb.R
\name{confint.cotframe}
\alias{confint.cotframe}
\title{Extract the confidence intervals from the the estimated changes over time measures}
\usage{
\method{confint}{cotframe}(object, parm = "coefficient", level = 0.95, ...)
}
\arguments{
\item{object}{a "cotframe"-class object}

\item{parm}{"coefficient". Confidence intervals are only available for AF measure point estimates.}

\item{level}{the confidence level required.}

\item{...}{additional argument(s) for methods.}
}
\value{
Confidence intervals extracted from the model \code{cotframe} object.
}
\description{
Extract the confidence intervals from the the estimated changes over time measures
}
\details{
The \code{confint} method for "cotframe"-class objects find the confidence
intervals from the changes over time estimates data frame. This method work for
only one measure \code{c("M0","H","A","hd","hdk")}. Then, user should subset the
data frame with the estimates by the chosen measure (including other preferred categories, i.e.,
poverty cut-off, subgroup, etc.)
}
\examples{
library(mpitbR)

data <- subset(syn_cdta, t==1)
data <- na.omit(data)

svydata <- survey::svydesign(id=~psu, weights = ~weight, strata = ~stratum, data = data)

indicators <- list(d1 = c("d_nutr","d_cm"),
                   d2 = c("d_satt","d_educ"),
                   d3 = c("d_elct","d_sani","d_wtr","d_hsg","d_ckfl","d_asst"))

set <- mpitb.set(svydata, indicators = indicators, name = "myname", desc = "pref. desc")

est <- mpitb.est(set, klist = c(33), measure = "M0",
                 weights = "equal", over = c("area"), multicore = TRUE)

coef(subset(est$lframe, measure == "M0"))
confint(subset(est$lframe, measure == "M0"))
summary(subset(est$lframe, measure == "M0"))
}
\seealso{
\code{coef}, and \code{summary} methods, and \code{mpitb.est} function.
}
\author{
Ignacio Girela
}
