% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.mpitb.R
\name{confint.lframe}
\alias{confint.lframe}
\title{Extract the confidence intervals from the the estimated cross-sectional measures}
\usage{
\method{confint}{lframe}(object, parm = "coefficient", level = 0.95, ...)
}
\arguments{
\item{object}{a "lframe"-class object}

\item{parm}{"coefficient". Confidence intervals are only available for AF measure point estimates.}

\item{level}{the confidence level required.}

\item{...}{additional argument(s) for methods.}
}
\value{
Confidence intervals extracted from the model \code{lframe} object.
}
\description{
Extract the confidence intervals from the the estimated cross-sectional measures
}
\details{
The \code{confint} method for "lframe"-class objects find the confidence
intervals from the different AF measures estimates data frame. This method work for
only one measure \code{c("M0","H","A","hd","hdk")} (Note that contribution measure
do no have confidence intervals). Then, user should subset the
data frame with the estimates by the chosen measure (including other preferred categories, i.e.,
poverty cut-off, subgroup, etc.)
}
\seealso{
\code{coef}, and \code{summary} methods, and \code{mpitb.est} function.
}
\author{
Ignacio Girela
}
