% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpitb.set.R
\name{mpitb.set}
\alias{mpitb.set}
\title{Set the specification of the Multidimensional Poverty Measurement and Analysis project}
\usage{
mpitb.set(data, indicators, ..., name = "unnamed", desc = "desc.")
}
\arguments{
\item{data}{a \code{survey.design}-class object where a complex survey design was previously specified. Can be a matrix but it is coerced to a `survey.design` class assuming equal probabilities.}

\item{indicators}{a list or character vector containing the names of the indicators. If it is a list, the element represents the dimension which character vector with their corresponding indicators. At most 10 dimensions are allowed. The indicators should belong to columns names of `data`.}

\item{...}{other arguments}

\item{name}{a character containing a desired specification of the project name. It also serves as an ID and it is recommended to use short names (at most 10 characters are permitted).}

\item{desc}{a character containing a desired specification of the project description.}
}
\value{
\code{mpitb_set}-class object
}
\description{
Set the specification of the Multidimensional Poverty Measurement and Analysis project
}
