% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mpitb.R
\name{print.summary.cotframe}
\alias{print.summary.cotframe}
\title{Print Summary function of summary.cotframe object}
\usage{
\method{print}{summary.cotframe}(
  x,
  digits = max(4, getOption("digits") - 3),
  signif.stars = getOption("show.signif.stars"),
  ...
)
}
\arguments{
\item{x}{the "summary.cotframe"-class object}

\item{digits}{controls number of digits printed in output.}

\item{signif.stars}{should significance stars be printed alongside output.}

\item{...}{optional arguments}
}
\value{
summarized object with nice format
}
\description{
Print Summary function of summary.cotframe object
}
