% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mpitb.R
\name{summary.cotframe}
\alias{summary.cotframe}
\title{Summary function of the estimates of the changes over time measures}
\usage{
\method{summary}{cotframe}(object, ...)
}
\arguments{
\item{object}{the "cotframe"-class object}

\item{...}{other arguments}
}
\value{
summary.cotframe a list of summary information for the estimated AF measure
}
\description{
Summary function of the estimates of the changes over time measures
}
\details{
The \code{summary} method for "cotframe"-class objects with coefficients, standard
errors and the corresponding t-values and the p-value for a two sided test,
\eqn{H_0: \beta = 0} being \eqn{\beta} any measure, with the 'significance stars'.
This method work for only one measure \code{c("M0","H","A","hd","hdk")} and either
relative or absolute measure, either non-annualized and annualized measure and only one
poverty cut-off.
Then, user should subset the data frame to obtain the desired summary.
}
\seealso{
\code{coef}, and \code{confint} methods, and \code{mpitb.est} function.
}
\author{
Ignacio Girela
}
