% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{syn_cdta}
\alias{syn_cdta}
\title{Synthetic dataset with a typical household survey design}
\format{
An object of class \code{data.frame} with 15000 rows and 17 columns.
}
\usage{
syn_cdta
}
\description{
This is a synthetic dataset drawn from the Stata mpitb package. This is the data used in the provided examples.
}
\references{
\doi{10.1177/1536867X231195286}
}
\author{
Nicolai Suppa
}
\keyword{datasets}
